Move business verification to the SmileIdentityCore class.
