+++ lib/IdApi.php
@@ -70,11 +70,7 @@ class IdApi
         $job_type = intval($partner_params["job_type"]);
 
         validateIdParams($id_info, $job_type);
-        validateJobTypes(array(JobType::ENHANCED_KYC, JobType::BUSINESS_VERIFICATION), $job_type);
-
-        if ($job_type === JobType::BUSINESS_VERIFICATION) {
-            return $this->submit_kyb_job($partner_params, $id_info);
-        }
+        validateJobTypes(array(JobType::ENHANCED_KYC), $job_type);
 
         $user_async = array_value_by_key("user_async", $options);
         $signature_params = $this->sig_class->generate_signature();
@@ -91,26 +87,6 @@ class IdApi
         $json_data = json_encode($data, JSON_PRETTY_PRINT);
         $url = $user_async ? 'async_id_verification' : 'id_verification';
         $resp = $this->client->post($url, [
-            'content-type' => 'application/json',
-            'body' => $json_data
-        ]);
-        $contents = $resp->getBody()->getContents();
-        return json_decode($contents, true);
-    }
-
-    private function submit_kyb_job($partner_params, $id_info)
-    {
-        $signature_params = $this->sig_class->generate_signature();
-        $data = array(
-            'partner_params' => $partner_params,
-            'partner_id' => $this->partner_id,
-            'source_sdk' => Config::SDK_CLIENT,
-            'source_sdk_version' => Config::VERSION
-        );
-        $data = array_merge($data, $id_info, $signature_params);
-        $json_data = json_encode($data, JSON_PRETTY_PRINT);
-
-        $resp = $this->client->post('business_verification', [
             'content-type' => 'application/json',
             'body' => $json_data
         ]);
