+++ tests/SmileIdentityCoreTest.php
@@ -382,4 +382,144 @@ final class SmileIdentityCoreTest extends TestCase
         $result = $this->sic->query_smile_id_services();
         $this->assertEquals($result, $expectedResult);
     }
+
+    public function testKybSuccessForBusinessRegistrationType()
+    {
+        $default_callback = 'https://google.com';
+        $api_key = file_get_contents(__DIR__ . "/assets/ApiKey.pub");
+        $sid_core = new SmileIdentityCore($this->partner_id, $default_callback, $api_key, $this->sid_server);
+
+        $partner_params = array(
+            'user_id' => '1',
+            'job_id' => '1',
+            'job_type' => JobType::BUSINESS_VERIFICATION
+        );
+
+        $id_info = array(
+            'country' => 'NG',
+            'id_type' => 'BUSINESS_REGISTRATION',
+            'id_number' => '00000000000',
+        );
+
+        $client = $this->getMockClient();
+        $sid_core->setClient($client);
+
+        $job = $sid_core->submit_job($partner_params, [], $id_info, []);
+        $this->assertEquals(array("success" => true), $job);
+    }
+
+    public function testKybSuccessForBasicBusinessRegistrationType()
+    {
+        $default_callback = 'https://google.com';
+        $api_key = file_get_contents(__DIR__ . "/assets/ApiKey.pub");
+        $sid_core = new SmileIdentityCore($this->partner_id, $default_callback, $api_key, $this->sid_server);
+
+        $partner_params = array(
+            'user_id' => '1',
+            'job_id' => '1',
+            'job_type' => JobType::BUSINESS_VERIFICATION
+        );
+
+        $id_info = array(
+            'country' => 'NG',
+            'id_type' => 'BASIC_BUSINESS_REGISTRATION',
+            'id_number' => '00000000000',
+        );
+
+        $client = $this->getMockClient();
+        $sid_core->setClient($client);
+
+        $job = $sid_core->submit_job($partner_params, [], $id_info, []);
+        $this->assertEquals(array("success" => true), $job);
+    }
+
+    public function testKybSuccessForTaxInformationType()
+    {
+        $default_callback = 'https://google.com';
+        $api_key = file_get_contents(__DIR__ . "/assets/ApiKey.pub");
+        $sid_core = new SmileIdentityCore($this->partner_id, $default_callback, $api_key, $this->sid_server);
+
+        $partner_params = array(
+            'user_id' => '1',
+            'job_id' => '1',
+            'job_type' => JobType::BUSINESS_VERIFICATION
+        );
+
+        $id_info = array(
+            'country' => 'NG',
+            'id_type' => 'TAX_INFORMATION',
+            'id_number' => '00000000000',
+        );
+
+        $client = $this->getMockClient();
+        $sid_core->setClient($client);
+
+        $job = $sid_core->submit_job($partner_params, [], $id_info, []);
+        $this->assertEquals(array("success" => true), $job);
+    }
+
+    public function testExceptionWhenJobTypeIsInvalid()
+    {
+        $expected_values = implode(", ", array(
+            JobType::BIOMETRIC_KYC,
+            JobType::DOCUMENT_VERIFICATION,
+            JobType::SMART_SELFIE_AUTHENTICATION,
+            JobType::BUSINESS_VERIFICATION
+        ));
+        $this->expectException(Exception::class);
+        $this->expectExceptionMessage("job_type must be one of $expected_values");
+        
+        $api_key = file_get_contents(__DIR__ . "/assets/ApiKey.pub");
+        $default_callback = 'https://google.com';
+        $sid_core = new SmileIdentityCore($this->partner_id, $default_callback, $api_key, $this->sid_server);
+
+        $partner_params = array(
+            'user_id' => '1',
+            'job_id' => '1',
+            'job_type' => 50 // invalid
+        );
+
+        $id_info = array(
+            'country' => 'NG',
+            'id_type' => 'BUSINESS_REGISTRATION',
+            'id_number' => '00000000000',
+        );
+
+        $sid_core->submit_job($partner_params, [], $id_info, []);
+    }
+
+    public function testInvalidIdTypeExceptionForKyb()
+    {
+        $expected_types = implode(", ", BusinessVerificationType::ALL);
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage("id_type must be one of $expected_types");
+        
+        $api_key = file_get_contents(__DIR__ . "/assets/ApiKey.pub");
+        $default_callback = 'https://google.com';
+        $sid_core = new SmileIdentityCore($this->partner_id, $default_callback, $api_key, $this->sid_server);
+
+        $partner_params = array(
+            'user_id' => '1',
+            'job_id' => '1',
+            'job_type' => JobType::BUSINESS_VERIFICATION
+        );
+
+        $id_info = array(
+            'country' => 'NG',
+            'id_type' => 'INVALID_TYPE',
+            'id_number' => '00000000000',
+        );
+
+        $sid_core->submit_job($partner_params, [], $id_info, []);
+    }
+
+    private function getMockClient()
+    {
+        $mock = new MockHandler([
+            new Response(200, [], '{"success":true}'),
+        ]);
+
+        $handler = HandlerStack::create($mock);
+        return new Client(['handler' => $handler]);
+    }
 }
